Add the fly to the title screen scene
Also, create two border walls. One set for normal operations, the second
for catching too many collisions at once. The fly uses a trigger
collider to check for "dangerous" objects (or just walls). And sort of
randomly decides where to go, based off of collision with a wall or with
an object.